+++ include/FFTProcessor.h
@@ -1,20 +1,24 @@
-#ifndef FFT_H
-#define FFT_H
+#ifndef FFTPROCESSOR_H
+#define FFTPROCESSOR_H
 
 #include <Arduino.h>
+#include <M5StickCPlus.h>
 #include <arduinoFFT.h>
 #include <driver/i2s.h>
 #include <math.h>
 
 class FFTProcessor
 {
+private:
 public:
     FFTProcessor();
-    ~FFTProcessor();
 
     bool setupI2Smic();
-    void I2SLoop();
     bool setupSpectrumAnalysis();
+    void setup();
+    void loop();
+
+    int *getLightness();
     bool getBeatHit();
 };
